+++ migration-engine/migration-engine-tests/tests/migrations/postgres.rs
@@ -104,7 +104,7 @@ async fn native_type_columns_can_be_created(api: &TestApi) -> TestResult {
     let types = &[
         ("smallint", "Int", "SmallInt", "int2"),
         ("int", "Int", "Integer", "int4"),
-        ("bigint", "Int", "BigInt", "int8"),
+        ("bigint", "BigInt", "BigInt", "int8"),
         ("decimal", "Decimal", "Decimal(4, 2)", "numeric"),
         ("numeric", "Decimal", "Numeric(4, 2)", "numeric"),
         ("real", "Float", "Real", "float4"),
